testing updated timer

+++ network.rb
@@ -30,7 +30,7 @@ $port = gets.chomp.to_i
 puts "Enter a percentage of packets to be dropped: "
 $pktpct = gets.chomp.to_i
 puts "Enter the delay in seconds (0.05) "
-$delay = gets.chomp.to_i
+$delay = gets.chomp.to_f
 
 network_1 = UDPSocket.new
 network_1.bind('', $port)
